requirements.txt
annotated-types==0.7.0
attrs==25.3.0
aws-cdk-lib==2.194.0
aws-cdk.asset-awscli-v1==2.2.235
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.cloud-assembly-schema==41.2.0
cattrs==24.1.3
certifi==2025.4.26
charset-normalizer==3.4.2
constructs==10.4.2
idna==3.10
importlib_resources==6.5.2
jsii==1.111.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.7
publication==0.0.3
pydantic==2.11.4
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
typeguard==2.13.3
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
